move components into components folder relates #20
